Picture retrieval code added and a bit of tweeking
So I've surrounded the functions that were there before in a document ready function so that they are run as soon as the page is loaded. So from now on any function call will have to be placed inside that thing.

Also I have attempted a getBackground function, I'm still struggling with retrieving the sizes of the images, I'll keep looking into that.

But the code for image retrieval is in getBackground()(src is the image). In order to change the topic of the image, in the url, where it says "&tags" we have to change the name (right now it's animals). I'm thinking of creating a small fucntion to randomize that.